move DeprecatedClass to widgets, where it's used
avoids need to add warn to genutils
