+++ IPython/html/widgets/widget_selection.py
@@ -2,17 +2,9 @@
 
 Represents an enumeration using a widget.
 """
-#-----------------------------------------------------------------------------
-# Copyright (c) 2013, the IPython Development Team.
-#
-# Distributed under the terms of the Modified BSD License.
-#
-# The full license is in the file COPYING.txt, distributed with this software.
-#-----------------------------------------------------------------------------
 
-#-----------------------------------------------------------------------------
-# Imports
-#-----------------------------------------------------------------------------
+# Copyright (c) IPython Development Team.
+# Distributed under the terms of the Modified BSD License.
 
 from collections import OrderedDict
 from threading import Lock
@@ -22,11 +14,9 @@ from IPython.utils.traitlets import (
     Unicode, Bool, Any, Dict, TraitError, CaselessStrEnum, Tuple, List
 )
 from IPython.utils.py3compat import unicode_type
-from IPython.utils.warn import DeprecatedClass
+from .deprecated import DeprecatedClass
 
-#-----------------------------------------------------------------------------
-# SelectionWidget
-#-----------------------------------------------------------------------------
+
 class _Selection(DOMWidget):
     """Base class for Selection widgets
